Traversing: move code to own module.
