+++ essence-of-live-coding/test/Cell.hs
@@ -25,6 +25,7 @@ import LiveCoding
 
 import qualified Cell.Monad.Trans
 import qualified Cell.Util
+import qualified Cell.Util.Traversable
 
 test =
   testGroup
@@ -36,6 +37,6 @@ test =
           sum (init inputs)
             === last (fst (runIdentity $ steps (sumC :: Cell Identity Int Int) inputs))
     , Cell.Util.test
-    , Cell.Util.testTraverse'
+    , Cell.Util.Traversable.testTraverse'
     , Cell.Monad.Trans.test
     ]
